requirements.txt
awsebcli==3.10.6
boto3==1.4.7
botocore==1.7.14
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.4
django-allauth==0.33.0
django-appconf==1.0.2
django-model-utils==3.0.0
django-notifications-hq==1.3
django-postman==3.5.1
django-storages==1.6.5
django-toolbelt==0.0.1
djangorestframework==3.6.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
jsonfield==2.0.2
oauthlib==2.0.3
olefile==0.44
pathspec==0.5.0
Pillow==4.2.1
psycopg2==2.7.1
python-dateutil==2.6.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
semantic-version==2.5.0
six==1.10.0
static3==0.7.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
websocket-client==0.44.0
